* Too many changes to note, mostly centered around Walls

+++ di/base/blocks/ISubItemRangeProvider.java
@@ -5,5 +5,5 @@ import com.google.common.collect.Range;
 
 public interface ISubItemRangeProvider 
 {
-	public HashMap<Short, Range> getRangeSet( );
+	public HashMap<Integer, Range> getRangeSet( );
 }
